requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
pyodbc==5.1.0
azure-storage-blob==12.19.0
python-dotenv==1.0.1
